requirements.txt
Django==1.9
olefile==0.44
Pillow==6.0.0
python-dotenv==0.13.0
python-http-client==3.2.7
sendgrid==6.2.2
